created components folder, working on migrating different components into their respective files
