+++ src/components/Square.js
@@ -1,0 +1,7 @@
+export default function Square({ value, onSquareClick }) {
+  return (
+    <button className="square" onClick={onSquareClick}>
+      {value}
+    </button>
+  );
+}
